Add files via upload
Day 2: I worked on the code once again, and I addressed a lot of bugs that I didn't notice on the first day. I also added a lot of comments so it's easier to understand the purpose and goal of each function. I still have a looping error that I can't seem to figure out what is causing the problem. Initially I thought it was because of the variable, but after some debugging, it doesn't seem to be the case. I may have to look towads the internet or office hours for help.