Add files via upload
I changed the way the function that tests if a number if prime number. Originally, it would divide the number being tested by every single other than 1 leading up to it. However, it now divides by every number that is already in the prime list and if it's divisible, it isn't prime and if it isn't, it returns True. Additionally, I addressed the initial issue that was causing the maximum recursion error, but it still returns maximum recursion error. I tried to add a line that increases the maximum recursion depth because that seems to be the only issue, but it hasn't worked yet.