Add files via upload
I worked on my code, but in terms of actually adding or editing lines, I didn't get much done. I was able to figure out why the recursion depth couldn't be changed and fixed it. I also figured out a couple of other errors. It took a really long time to run the code twice, and both times, I got a stack overflow. I noticed the problem was that it was for some reason passing every number as a prime number which increase the run time significantly, so I will work on that the next work session