Add files via upload
I was not able to change a lot of the actual program this time around. It took me a while to figure out the program was continuously returning true for numbers that weren't prime numbers. However, I was able to figure out what was wrong after deleting a bit of code and using Python Tutor to visualize the program. In the end, the programming is working as anticipated except it is returning system overflow. I need to figure out a way to get around this or use an alternative method that I can optimize the program. Worse comes to worse, I think I will end up changing the is_prime program to the module that has the program built in.